More factoring and includes
