+++ src/comms/hid_comms.cpp
@@ -1,3 +1,4 @@
+#include <Arduino.h>
 #include "hid_comms.hpp"
 
 namespace Comms {
